pyproject.toml
[project]
name = "planner-app"
version = "0.1.0"
description = "Weekly planner (Mon–Sun), PostgreSQL + Flask"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "Flask>=3.0.0",
    "psycopg2-binary>=2.9.9",
]

[tool.uv]
dev-dependencies = []
